third_party/WebKit/LayoutTests/accessibility/main-element-expected.txt

main

This tests that the HTML5 main element correctly maps to AXMain.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS main.role is 'AXRole: AXMain'
PASS successfullyParsed is true

TEST COMPLETE
